+++ Source/DTCollectionViewManager+Drag.swift
@@ -29,7 +29,7 @@ import DTModelStorage
 
 extension DTCollectionViewManager {
     
-    #if os(iOS) && swift(>=3.2)
+    #if os(iOS)
     
     @available(iOS 11, *)
     /// Registers `closure` to be executed when `UICollectionViewDragDelegate.collectionView(_:itemsForBeginning:at:)` method is called for `cellClass`.